Change how odf2jats is invoked, now specifying only the folder: ./odf2jats source/odf-bringedal-editorial odf2jats then calls calabash with folder/content.xml as source and folder/styles.xml as a parameter named documentStylesPath
This is to prepare to extract text from headers and footers that are stored in the
styles.xml document.

Also try to improve detection of styles of headings and other elements,
but it is possible that some regressions have been introduced.

Needs more testing.

Also needs to extract text from the header, which means to extract
and transform text:p elements in styles.xml to appropriate elements