refactor the pipeline a bit and change the commandline to call the pipeline
now using both parameters (to be passed to p:xslt)

and options given on the commandline to be accessed by the XProc pipeline.

options are both passed on the commandline to odf2jats.xpl and need to be declared as <p:option name="optionName" required="true"/> elements in odf2jats.xpl
then they can be accessed as any variable... inside odf2jats.xpl select="$optionName"

Parameters do not need to be known to the XProc pipeline, they are silently being passed to steps that are expecting them.

The xsl-stylesheets using those parameters will have to decalare them, such as in flat_odf_extract.xsl where the stylepath parameter is declared:
<xsl:param name="stylepath"/>

now the xproc pipeline will create and store the result xml into a file, to a computed path by using options for path and filenames,
passed to the pipeline, were before this was handled by redirecting standard input to file with the tee command